One more revision of the email system.
It's now a dovecot+mbsync+postfix system.

The idea being that mbsync fetches the mails from the various imap
servers and sync it with a local dovecot imap server. That local dovecot
server can then be accessed by any (local) email client, like Airmail
(tested) or mutt (not yet tested in this mode of operation). Postfix is
used to relay the outgoing mails to the relevant relays. Authentication
to the local dovecot (auth shared by postfix) is made by local-only
passwords that are generated and stored in the mac keychain (and
read/written with the security mac command line utility).

The postfix passwords db for the relays is also generated from the mac
keychain, assuming we've entered there the real passwords for the mail
relays in mbsync-$account-password entries.